# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API Portal de compras - Farmácias Associadas",
        "version": "1.0.0"
    },
    "paths": {
...
        "/offers": {
            "get": {
                "tags": [
                    "Offers"
                ],
                "parameters": [
                    {
                        "name": "name",
                        "in": "query",
                        "example": "teste"
                    },
                    {
                        "name": "code",
                        "in": "query",
                        "example": "123"
                    },
                    {
                        "name": "status",
                        "in": "query",
                        "example": "active"
                    },
                    {
                        "name": "startDate",
                        "in": "query",
                        "example": "2020-05-25 03:00:00"
                    },
                    {
                        "name": "endDate",
                        "in": "query",
                        "example": "2020-06-25 03:00:00"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/OfferListResource"
                                            }
                                        },
                                        "links": {
                                            "allOf": [
                                                {
                                                    "$ref": "#/components/schemas/PaginationLinks"
                                                }
                                            ]
                                        },
                                        "meta": {
                                            "allOf": [
                                                {
                                                    "$ref": "#/components/schemas/PaginationMeta"
                                                }
                                            ]
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Offers"
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/OfferCreatorRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "example": "Offera criado com sucesso"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ValidationResponse"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "example": "Mensagem de erro"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/offers/{id}": {
            "get": {
                "tags": [
                    "Offers"
                ],
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "example": "2"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "$ref": "#/components/schemas/OfferListResource"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "example": "Mensagem de error"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "put": {
                "tags": [
                    "Offers"
                ],
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "example": "2"
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/OfferUpdatorRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "example": "Offera atualizado com sucesso"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ValidationResponse"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "example": "Mensagem de erro"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "delete": {
                "tags": [
                    "Offers"
                ],
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "example": "2"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "example": "Offera removido com sucesso"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "example": "Mensagem de error"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "OfferCreatorRequest": {
                "title": "Offer form request",
                "required": [
                    "code",
                    "name",
                    "status",
                    "description"
                ],
                "properties": {
                    "code": {
                        "type": "integer",
                        "example": "001"
                    },
                    "name": {
                        "type": "string",
                        "example": "Teste"
                    },
                    "startDate": {
                        "type": "datetime",
                        "example": "2020-05-25T05:09:15.000000Z"
                    },
                    "endDate": {
                        "type": "datetime",
                        "example": "2020-05-25T05:09:15.000000Z"
                    },
                    "condition": {
                        "type": "string",
                        "example": "name"
                    },
                    "minimumPrice": {
                        "type": "string",
                        "example": "500"
                    },
                    "offerType": {
                        "type": "string",
                        "example": "null"
                    },
                    "sendType": {
                        "type": "string",
                        "example": "null"
                    },
                    "noAutomaticSending": {
                        "type": "boolean",
                        "example": "true"
                    },
                    "impound": {
                        "type": "boolean",
                        "example": "false"
                    },
                    "description": {
                        "type": "string",
                        "example": "asdasd"
                    },
                    "emails": {
                        "$ref": "#/components/schemas/Email"
                    },
                    "status": {
                        "$ref": "#/components/schemas/UserStatus"
                    }
                },
                "type": "object"
            },
            "OfferUpdatorRequest": {
                "title": "Offer update form request",
                "properties": {
                    "code": {
                        "type": "integer",
                        "example": "001"
                    },
                    "name": {
                        "type": "string",
                        "example": "Teste"
                    },
                    "startDate": {
                        "type": "datetime",
                        "example": "2020-05-25T05:09:15.000000Z"
                    },
                    "endDate": {
                        "type": "datetime",
                        "example": "2020-05-25T05:09:15.000000Z"
                    },
                    "condition": {
                        "type": "string",
                        "example": "name"
                    },
                    "minimumPrice": {
                        "type": "string",
                        "example": "500"
                    },
                    "offerType": {
                        "type": "string",
                        "example": "null"
                    },
                    "sendType": {
                        "type": "string",
                        "example": "null"
                    },
                    "noAutomaticSending": {
                        "type": "boolean",
                        "example": "true"
                    },
                    "impound": {
                        "type": "boolean",
                        "example": "false"
                    },
                    "description": {
                        "type": "string",
                        "example": "asdasd"
                    },
                    "emails": {
                        "$ref": "#/components/schemas/Email"
                    }
                },
                "type": "object"
            },
...
            "DistributorListResource": {
                "title": "Distributor Response",
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": "1"
                    },
                    "code": {
                        "type": "string",
                        "example": "01"
                    },
                    "cnpj": {
                        "type": "string",
                        "example": "00.00001"
                    },
                    "name": {
                        "type": "string",
                        "example": "Teste"
                    },
                    "status": {
                        "$ref": "#/components/schemas/UserStatus"
                    }
                },
                "type": "object"
            },
...
            "OfferListResource": {
                "title": "Offer List Response",
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": "1"
                    },
                    "code": {
                        "type": "string",
                        "example": "01"
                    },
                    "name": {
                        "type": "string",
                        "example": "Teste"
                    },
                    "startDate": {
                        "type": "datetime",
                        "example": "2019-05-25 03:00:00"
                    },
                    "endDate": {
                        "type": "datetime",
                        "example": "2019-07-25 03:00:00"
                    },
                    "condition": {
                        "type": "string",
                        "example": "teste"
                    },
                    "minimumPrice": {
                        "type": "string",
                        "example": "500"
                    },
                    "offerType": {
                        "type": "string",
                        "example": "null"
                    },
                    "sendType": {
                        "type": "string",
                        "example": "null"
                    },
                    "noAutomaticSending": {
                        "type": "boolean",
                        "example": "true"
                    },
                    "impound": {
                        "type": "boolean",
                        "example": "false"
                    },
                    "description": {
                        "type": "string",
                        "example": "asdqwe"
                    },
                    "email": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/Email"
                        }
                    },
                    "partners": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/DistributorListResource"
                        }
                    }
                },
                "type": "object"
            },
...
            "UserStatus": {
                "type": "string",
                "enum": [
                    "ACTIVE",
                    "INACTIVE"
                ]
            },
...
            "PaginationLinks": {
                "properties": {
                    "first": {
                        "type": "string",
                        "example": "http://api.test/api/users?page=1"
                    },
                    "last": {
                        "type": "string",
                        "example": "http://api.test/api/users?page=10"
                    },
                    "next": {
                        "type": "string",
                        "example": "http://api.test/api/users?page=3"
                    }
                },
                "type": "object"
            },
...
            "Email": {
                "properties": {
                    "send": {
                        "type": "string",
                        "example": "TO"
                    },
                    "email": {
                        "type": "string",
                        "example": "teste@gmail.com"
                    }
                },
                "type": "object"
            },
...
            "PaginationMeta": {
                "properties": {
                    "current_page": {
                        "type": "integer",
                        "example": "2"
                    },
                    "from": {
                        "type": "integer",
                        "example": "1"
                    },
                    "last_page": {
                        "type": "integer",
                        "example": "10"
                    },
                    "path": {
                        "type": "string",
                        "example": "http://api.test/api/users"
                    },
                    "per_page": {
                        "type": "integer",
                        "example": "20"
                    },
                    "to": {
                        "type": "integer",
                        "example": "3"
                    },
                    "total": {
                        "type": "integer",
                        "example": "10"
                    }
                },
                "type": "object"
            },
            "ValidationResponse": {
                "properties": {
                    "message": {
                        "type": "string",
                        "example": "The given data was invalid."
                    },
                    "errors": {
                        "type": "array",
                        "items": {
                            "type": "array",
                            "items": {
                                "type": "string"
                            },
                            "example": "Campo teste é obrigatório"
                        }
                    }
                },
                "type": "object"
            },
...
        },
...
        "securitySchemes": {
            "passport": {
                "type": "oauth2",
                "description": "Laravel passport oauth2 security.",
                "in": "header",
                "scheme": "https",
                "flows": {
                    "password": {
                        "authorizationUrl": "http://localhost/oauth/authorize",
                        "tokenUrl": "http://localhost/oauth/token",
                        "refreshUrl": "http://localhost/token/refresh",
                        "scopes": []
                    }
                }
            }
        }
    }
}